requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
django==5.2 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.16.0 ; python_version >= "3.10" and python_version < "4.0"
graphql-core==3.2.6 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.10 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
strawberry-graphql-django==0.58.0 ; python_version >= "3.10" and python_version < "4.0"
strawberry-graphql==0.263.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
